==> notebooks/06.1.businesses.by.zip.ipynb <==

import pandas as pd
import requests
import os, sys, inspect
import numpy as np
import matplotlib.pyplot as plt
# add parent dir to system dir
currdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
rootdir = os.path.dirname(currdir)
sys.path.insert(0, rootdir)

from src.config import googleapikey

ffname = os.path.join(rootdir, "data", "int", "zip_codes.csv")
biz_by_zip_df = pd.read_csv(ffname, header=None, names=["Zip Code"])

biz_by_zip_df.head()

#adding lat/lon columns for google search
biz_by_zip_df["Lat"] = ""
biz_by_zip_df["Lng"] = ""

# set up params
params = {
    "key": googleapikey
}

# Loop through and grab the lat/lng using Google maps
for index, row in biz_by_zip_df.iterrows():

    # update params with zipcode each loop
    params['address'] = row['Zip Code']

    base_url = "https://maps.googleapis.com/maps/api/geocode/json"

    # Run requests to grab the JSON at the requested URL
    zip_location = requests.get(base_url, params=params)
    print(zip_location.url)

    zip_location = zip_location.json()

    # Append the lat/lng to the appropriate columns
    # Use try / except to skip any cities with errors
    try:
        zip_lat = zip_location["results"][0]["geometry"]["location"]["lat"]
        zip_lng = zip_location["results"][0]["geometry"]["location"]["lng"]

        biz_by_zip_df.loc[index, "Lat"] = zip_lat
        biz_by_zip_df.loc[index, "Lng"] = zip_lng

    except (KeyError, IndexError):
        print("Error with city data. Skipping")

        # drops rows with errors
        biz_by_zip_df.drop(labels=index, inplace=True)


# Visualize
biz_by_zip_df.head()

Bank Count

# Create an empty column for bank count
biz_by_zip_df["Bank Count"] = ""

# Set up params
params = {
    "radius": 5000,
    "type": "bank",
    "key": googleapikey
}

# Loop through and run Google search to get all banks in 3.1 mile radius (5000 meters)
for index, row in biz_by_zip_df.iterrows():

    # Get lat/lng from df
    lat = row['Lat']
    lng = row['Lng']

    # Update params
    params['location'] = f"{lat},{lng}"

    # Create endpoint url using Google Places Radar and the lat/lng we identified earlier
    base_url = "https://maps.googleapis.com/maps/api/place/radarsearch/json"

    # Run a request, print the url, and convert to json
    bank_data = requests.get(base_url, params=params)
    print(bank_data.url)
    bank_data = bank_data.json()

    # Measure bank count on the number of results in the retrieved area
    bank_count = len(bank_data["results"])

    print(f"Final Bank Count: {bank_count}")
    print("")

    # Store the bank count into the Data Frame
    biz_by_zip_df.loc[index, "Bank Count"] = bank_count

    # Reset bank_count (so there is no chance that a previous record is influencing a latter one)
    bank_count = 0

# Visualize
biz_by_zip_df.head()

Liquor Store Count

# Create an empty column for liquor store count
biz_by_zip_df["Liquor Store Count"] = ""

# Set up params
params = {
    "radius": 5000,
    "type": "liquor_store",
    "key": googleapikey
}

# Loop through and run Google search to get all liquor stores in 3.1 mile radius (5000 meters)
for index, row in biz_by_zip_df.iterrows():

    # Get lat/lng from df
    lat = row['Lat']
    lng = row['Lng']

    # Update params
    params['location'] = f"{lat},{lng}"

    # Create endpoint url using Google Places Radar and the lat/lng we identified earlier
    base_url = "https://maps.googleapis.com/maps/api/place/radarsearch/json"

    # Run a request, print the url, and convert to json
    liquor_data = requests.get(base_url, params=params)
    print(liquor_data.url)
    liquor_data = liquor_data.json()

    # Measure liquor store count on the number of results in the retrieved area
    liquor_count = len(liquor_data["results"])

    print(f"Final Liquor Store Count: {liquor_count}")
    print("")

    # Store the liquor store count into the Data Frame
    biz_by_zip_df.loc[index, "Liquor Store Count"] = liquor_count

    # Reset liquor_count (so there is no chance that a previous record is influencing a latter one)
    liquor_count = 0

# Visualize
biz_by_zip_df.head()

Wal-Mart Count

# Create an empty column for walmart count
biz_by_zip_df["Wal-Mart Count"] = ""

# Set up params
params = {
    "radius": 5000,
    "name": "Wal-Mart",
    "key": googleapikey
}

# Loop through and run Google search to get all walmarts in 3.1 mile radius (5000 meters)
for index, row in biz_by_zip_df.iterrows():

    # Get lat/lng from df
    lat = row['Lat']
    lng = row['Lng']

    # Update params
    params['location'] = f"{lat},{lng}"

    # Create endpoint url using Google Places Radar and the lat/lng we identified earlier
    base_url = "https://maps.googleapis.com/maps/api/place/radarsearch/json"

    # Run a request, print the url, and convert to json
    walmart_data = requests.get(base_url, params=params)
    print(walmart_data.url)
    walmart_data = walmart_data.json()

    # Measure walmart count on the number of results in the retrieved area
    walmart_count = len(walmart_data["results"])

    print(f"Final Wal-Mart Count: {walmart_count}")
    print("")

    # Store the walmart count into the Data Frame
    biz_by_zip_df.loc[index, "Wal-Mart Count"] = walmart_count

    # Reset walmart_count (so there is no chance that a previous record is influencing a latter one)
    walmart_count = 0

# Visualize
biz_by_zip_df.head()

Target Count

# Create an empty column for target count
biz_by_zip_df["Target Count"] = ""

# Set up params
params = {
    "radius": 5000,
    "name": "Target",
    "key": googleapikey
}

# Loop through and run Google search to get all Targets in 3.1 mile radius (5000 meters)
for index, row in biz_by_zip_df.iterrows():

    # Get lat/lng from df
    lat = row['Lat']
    lng = row['Lng']

    # Update params
    params['location'] = f"{lat},{lng}"

    # Create endpoint url using Google Places Radar and the lat/lng we identified earlier
    base_url = "https://maps.googleapis.com/maps/api/place/radarsearch/json"

    # Run a request, print the url, and convert to json
    target_data = requests.get(base_url, params=params)
    print(target_data.url)
    target_data = target_data.json()

    # Measure target count on the number of results in the retrieved area
    target_count = len(target_data["results"])

    print(f"Final Target Count: {target_count}")
    print("")

    # Store the target count into the Data Frame
    biz_by_zip_df.loc[index, "Target Count"] = target_count

    # Reset target_count (so there is no chance that a previous record is influencing a latter one)
    target_count = 0

# Visualize
biz_by_zip_df.head()

Starbucks Count

# Create an empty column for Starbucks count
biz_by_zip_df["Starbucks Count"] = ""

# Set up params
params = {
    "radius": 5000,
    "name": "Starbucks",
    "key": googleapikey
}

# Loop through and run Google search to get all Starbucks in 3.1 mile radius (5000 meters)
for index, row in biz_by_zip_df.iterrows():

    # Get lat/lng from df
    lat = row['Lat']
    lng = row['Lng']

    # Update params
    params['location'] = f"{lat},{lng}"

    # Create endpoint url using Google Places Radar and the lat/lng we identified earlier
    base_url = "https://maps.googleapis.com/maps/api/place/radarsearch/json"

    # Run a request, print the url, and convert to json
    starbucks_data = requests.get(base_url, params=params)
    print(starbucks_data.url)
    starbucks_data = starbucks_data.json()

    # Measure Starbucks count on the number of results in the retrieved area
    starbucks_count = len(starbucks_data["results"])

    print(f"Final Starbucks Count: {starbucks_count}")
    print("")

    # Store the Starbucks count into the Data Frame
    biz_by_zip_df.loc[index, "Starbucks Count"] = starbucks_count

    # Reset starbucks_count (so there is no chance that a previous record is influencing a latter one)
    starbucks_count = 0

# Visualize
biz_by_zip_df.head()

output = os.path.join(rootdir, "data", "int", "businesses_by_zip.csv")
biz_by_zip_df.to_csv(output, index=False)
biz_by_zip_df
