==> notebooks/10.2.heat.map.by.year.month.ipynb <==
import pandas as pd
import numpy as np
import os, sys, inspect
import folium
import subprocess
from folium.plugins import HeatMap
from sklearn import preprocessing

# add parent dir to system dir
currdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
rootdir = os.path.dirname(currdir)
sys.path.insert(0, rootdir)

from src.utils import get_temporal_data, get_spatial_data

ffname_realestate = os.path.join(rootdir, "data", "int", "realestate_demographics.csv")
realestate_df = pd.read_csv(ffname_realestate)
#realestate_df.head()

# ziptrend = get_temporal_data(78701,realestate_df)

# ziptrend.head()

fname = "Home Value Index"
yeartrend = get_spatial_data(2017,realestate_df)
yeartrend12 = yeartrend.loc[yeartrend["Month"]==12,:]
#yeartrend12.head()


column_to_heatmap = "Home Value Index"
yeartrend_df = yeartrend12[["Latitude","Longitude",column_to_heatmap]]


yeartrend_df.reset_index(inplace=True)
yeartrend_df

yeartrend_df.drop(["index"], inplace=True, axis=1)
yeartrend_df

data = yeartrend_df.values

minmax_scalar = preprocessing.MinMaxScaler()
data[:,[2]] = minmax_scalar.fit_transform(data[:,[2]])
data

# heatmap params

location = [30.3272, -97.7431]
zoom_start = 11
name = column_to_heatmap
radius = 30

# create map
m = folium.Map(location, zoom_start=zoom_start)
HeatMap(data.tolist(),radius=radius,min_opacity = .8, blur = 45, gradient = {0.4: 'blue', 0.65: 'lime', 1: 'red'}).add_to(m)
m.save(os.path.join(rootdir, 'Reports', 'Figures', (name + '.html')))

help(HeatMap)

help(folium.Map)
